# Optimize Module#delegate to no longer be linear